api/src/core/auth/local.strategy.ts
import { Strategy } from 'passport-local';
import { PassportStrategy } from '@nestjs/passport';
import { Injectable, UnauthorizedException } from '@nestjs/common';
import { AuthService } from './auth.service';
import { IUser } from '../types';

@Injectable()
export class LocalStrategy extends PassportStrategy(Strategy) {
  constructor(private auth: AuthService) {
    super();
  }

  /**
   * Validate provided credentials
   *
   * @param {string} email
   * @param {string} password
   *
   * @returns {Promise<Partial<IUser>>}
   */
  async validate(email: string, password: string): Promise<Partial<IUser>> {
    const user = await this.auth.validateUser(email, password);
    if (!user) {
      throw new UnauthorizedException();
    }
    return user;
  }
}
